Fix hhbbc performance regression
Summary:
Function folding, and context sensitive return types depend on hashing
the arguments. Type::hash was very simple though, not taking much
about the type into account.

If a single function was called from a lot of different places, with
similar argument types (but different values), we could hit a lot of
contention.

Reviewed By: oulgen

Differential Revision: D15425083

fbshipit-source-id: 35c7ba110f1a1b37951fc8aaa09beac87c7de0e2