Add a runtime option to control the amount of hot code placed on huge pages
Summary:
All the hot code section was being placed on huge pages, with no knob to disable
it.  This diff adds the ability to control how much hot code is placed on huge
pages.  This diff changes the meaning of the TCNumHugeHotMB option to mean the
hot code area instead of the main one, and adds a TCNumHugeMainMB to control the
amount of huge pages used by the main area.

Reviewed By: binliu19

Differential Revision: D15422922

fbshipit-source-id: 4f7d71ab36e7737c79a963384967fafbf8825cd4