Carry async information into heap
Summary:
The runtime enforces (enforceable) inner types of Awaitables provided that a function is declared async. Function types are stored in the heap, but they omit this information. This adds the information to the heap via the fun_type record in Typing_defs.

I put in a dummy FSync for lambdas with a TODO. The problem is that anonymous functions don't carry their async declaration into the environment; for now it's OK to be more pessimisitc about their return types than for declared functions.

Reviewed By: Matt-Schellhas

Differential Revision: D15420320

fbshipit-source-id: 31705348e7cfb4b2c0e49ccc392646903dd172f6