Convert AsTypeStruct to always throw
Summary:
Originally `$x as T` was intended to be written as
```
push $x
push TS for T
AsTypeStructC
```
but we wanted it to be faster when it does not fail so wee converted it to
```
($x is T) ? $x : ($x as T)
```
This diff converts `AsTypeStructC` instruction to `ThrowAsTypeStructException` since if we are running `AsTypeStructC` we know that it will fail but the JIT/hhbbc doesn't. This will enable us to emit better code.

Reviewed By: markw65

Differential Revision: D15402725

fbshipit-source-id: 0a0da3336bcebb0613676f989e523f7ec0f0e7f4