Check for invalid type hint to is/as expressions even when we do not resolve the TS
Summary: When we use is/as expressions with reified types, we use DontResolve flag on IsTypeStruct because reified types are already resolved so we do not need to pay the cost of re-resolving them. However, this was leading us to not detect the invalid types for is/as expressions like callables etc. These invalid types are okay to reify but not okay to use with is/as expressions.

Reviewed By: markw65

Differential Revision: D15426340

fbshipit-source-id: 67f69bc1bb1287ac7b08bc2109d29e9052e8104b