Make Shapes::idx an optimized builtin
Summary:
Array-offset profiling doesn't work on shapes, because we don't inline Shapes::idx in profiled translations, and so we mix together profiles across different key accesses.

This diff is an attempt to fix the issue by making Shapes::idx a native builtin (thanks Rick Lavoie for the suggestion). Specifically, it:
1. Moves the Shapes class to an extension library and rewrites Shapes::idx as a native static method.
2. Adds an optimized path when using FCallBuiltin with Shapes::idx.

I've manually verified that the optimized path with FCallBuiltin generates the right opcodes:
- When profiling, ProfileMixedArrayOffset + ArrayIdx
- When optimizing, CheckMixedArrayOffset + MixedArrayGetK

The test case uses two different keys with Shapes::idx, so it's an example where array offset profiling fails before.

Reviewed By: mxw

Differential Revision: D15295522

fbshipit-source-id: 6995df09a01e67ae244257711ada7a4b3e7293bf