Removed autoloader kind for records
Summary:
Records now use the same autoloader map as classes. With this change, there will be no change required in the www code to support autoloading of records. We still need to support records in facts parser, but it is not part of this diff.

The function `getStringRepr` returns `class` for both classes (including trats, interfaces and enums) and records. This can be confusing, but changing this string is not straightforward as it requires changes in www too. As Jeff i working on native autoloader support, he can change the string when the www autoloader is not used anymore.

Reviewed By: jthemphill

Differential Revision: D15153179

fbshipit-source-id: a0a13a9ba4d4a173ac80663351b1ed1eb050ebf5