Set up separate process
Summary: For robustness, the serverless IDE should run in a separate process. It will help to avoid blocking the main thread with CPU-bound work (particularly parsing and typechecking, in pathological cases), and will let the OOM-killer kill only the IDE services without taking down the entire language server (although arguably in that case, we should recover by just crashing the entire language server -- we can do that later if we so decide).

Reviewed By: 2BitSalute

Differential Revision: D15252980

fbshipit-source-id: d0af8af80ab64245e7afa37359490c58f7db6ab6