Clean up resources on shutdown
Summary:
Allows us to clean up resources on shut down. We accomplish this by sending a `Shutdown` RPC message to the serverless IDE process when we're asked to shut down.

Note that this is a best-effort cleanup. It is possible for someone to unceremoniously kill our processes. If there's something seriously important which needs to be cleaned up, it should be done transactionally.

The Language Server Protocol defines two methods for instructing the language server to shut down:

1) The `shutdown` request. The language client should wait for the language server to do its shutdown and finally respond to the `shutdown` request. This way, the language client knows that it's safe to actually kill the process.
2) The `exit` notification. This does not have a corresponding response sent by the language server. It's expected that the language server will promptly shut itself down at this stage. The language client could reasonably kill it at any time after issuing the `exit` notification.

We do our cleanup in the first one, since it advises the language client to hold off on actually killing the process.

Reviewed By: 2BitSalute

Differential Revision: D15252985

fbshipit-source-id: 8d70d0f9d8e3ad3881d5904eece77d5e6c2662bc