Fix static class property/method lookup in unions
Summary: At the moment, when we look up a static method `f` in `$c` (using `$c::f`) where `$c` has a union type, say `C1 | C2`, we separately look up `f` in `C1` and `C2`, and then localize the corresponding function types using the same expansion environment `ety_env` in which `this_ty` is set to the union `C1 | C2`.  If the type of `f` in `C2` contains references to type constants defined in `C2`, we will try to look up those type constants in the union and hence also in `C1`, in which they might not even be defined.  This will result in a very confusing error which, furthermore, has the wrong primary position, pointing to the definition of `C1` not the expression `$c::f`.  To fix this, we need to set `this_ty` appropriately for each element of the union.

Reviewed By: dlreeves

Differential Revision: D15449001

fbshipit-source-id: 62ea6009a56c638a374711f823926569bff4ab5d