Add ProvenanceSkipFrame attr
Summary:
What it says on the tin.

When in the context of a function with `__ProvenanceSkipFrame`, operations that
would have assigned a provenance tag... won't. Static arrays that would
normally be attributed to these functions ... won't.

We'll need to work around this somehow (probably, by forcing a copy of the
previously-static array to be made at some point in control flow?)

Alternately, we can have static arrays in these extents be marked with a
provenance tag that is eagerly overwritten?

Reviewed By: mxw

Differential Revision: D15031430

fbshipit-source-id: a2de7922d774f96896ba2e697d12c5451fc76207