Test Facts Parser as used by hh_single_compile
Summary:
Add the missing test coverage of FactParser  by simply
comparing its `extract_to_json` outputs from both OCaml and Rust
(after normalizing the JSONs via pretty printing and sorting)
on the entire test set of Hack (or HHVM if specified) files.
Highlight the differences in JSON to aid in fixing inconsistencies
unless `--no-diff` is shown (useful with `--keep-going`).

This is written as a pure I/O test, not a part of `rust_ocaml`
because it would be awkward to reimplement yet another serializer
of facts (just so that it can be nicely compared in OCaml) or
writing sophisticated to-OCaml serialization of recursive Node type
in FactsSmartConstructors. It is sufficient as serialization from
Facts is *already* covered by unit tests in `src/facts:facts_rust`.

Reviewed By: Wilfred

Differential Revision: D15407860

fbshipit-source-id: 91b03ad3b9c130721299d32a8ca430890e488aa2