Do not re-generate Rust files by default, require opt-in with a flag
Summary:
Leo, Steve: take note of this diff, because it makes us have to re-run the generator manually and introduces the possibility of segfaults until we do.

Quoting the comment:

"One of the promises of Rust migration is that we will not force people to
write the same code twice as long as possible. To achieve that, all the
tests testing Rust codepaths are with "disabled" tag and require explicit
--run-disabled invocation. But it's not enough - even if this allows
others to land code that breaks those tests (so we can fix it later for
them), the code still needs to compile.

This means that changing schema (by renaming / removing items) and
re-running generator can force people to update Rust code too (and writing
the same code twice), which leaves a bad first impression and can deplete
people's goodwill banks for this project.

As a mitigation, we won't regenerate Rust code by default, unless you run
the generator with --rust flag. That means that the parsers can diverge,
in which case calling Rust from OCaml can now lead to segfaults /
unexpected results until the code is regenerated. Effectively, we are
moving the possibly bad experience from "everyone" to "people working on
Rust port"."

Initially I forgot about the compilation part and thought that disabled tests is enough, but recently had multiple people run into it. Some of them just silently updated the Rust version (Kendall, Vassil), others were surprised and had to reach out to us for clarification (Xilun, Arnab, Alexey). Let's try to remove this annoyance for now.

Reviewed By: losvald

Differential Revision: D15486128

fbshipit-source-id: 525e16d586fccfbac79ecc66782238ca702a7821