Clear top function on fpiStack() in region-tracelet
Summary:
When generating live translations we include the specialized types of objects in the RegionContext during tracelet selection, but if no one elects to constrain those types we relax them away prior to generating the live translation. The simplifier inspects the specialized types of objects passed to `LdObjClass` and as a result we sometimes push known callees to the stack during region selection even when the closure type is going to be relaxed.

This hadn't been a problem before as the region with and without this known callee information was mostly the same, however a recent change (D15325174) allowed us to extract the callee from the `fpiStack()` on FrameState even in cases where we believed the callee was unknown. In this particular instance, we would then sometimes infer a return type during region selection preventing us from breaking the tracelet on the first bytecode to consume the output of the function call. When generating the final translation we would then drop the inferred type along with the callee (when the closure was unspecialized), leading do malformed bytecode (SetM and many other bytecodes depend on their stack and local inputs being known DataTypes).

Reviewed By: mxw

Differential Revision: D15493575

fbshipit-source-id: 6eba67fadd75b238bf76a7fc7e2b931d7b61c9db