Add test suite for autocomplete
Summary:
Add a basic test for autocomplete.

Plan for future diffs in this stack to introduce more functionality and more tests.  As per discussion with Waleed Khan what will be really interesting is when we build an index, test it, modify a few files, and test it again.  But that will be the next diff in the stack.

Reviewed By: 2BitSalute

Differential Revision: D15403579

fbshipit-source-id: bb12fd886614d229d7e52ceb730cf93c02d6c54a