Filter sqlite results by tombstones
Summary:
Enable the "local search service" to detect when a file has been deleted locally.

Note that this simply removes dead answers; it does not automatically bring in additional answers if the results have been filtered.  So if you retrieve 100 results from the global index, and five of them are from files that have been removed locally, you will end up with 95 results from the global index.  I'm not yet certain how big of a problem this is in practice; if necessary I will take additional steps to address filtering.

== initial joke summary below this line ==

Remember when we could move a file from one location to another, autocomplete search for it, and not see duplicate results?

{F158833918}

Remember when we could remove a file, autocomplete search for it, and see nothing?

{F158834860}

Reviewed By: 2BitSalute

Differential Revision: D15363771

fbshipit-source-id: c20316c5ee991d791c246797ed75ef4779f9e84e