Make declared function return type information available to type checking of `return`
Summary:
The type checking of the `return` statement depends on a data structure called `Typing_env_return_info.t`, which collects the current function's return type (`return_type`) and a few flags. However, this return type is a `locl ty`, so it lacks information about its origin. For example, if we had `return_type: locl ty = Tprim Nast.Tint`, we can't readily determine whether that came from
```
function f(): int { return 4; }
```
or
```
class C {
  const type T = int
  public function g(): this::T { return 4; }
}
```

It is an important distinction because HHVM enforces the first return type but not the second. The type checking of `return` hands the return type to `Typing_coercion` which checks to see whether the return type can coerce into the return type or is otherwise a subtype.

In order to check whether this return type can be trusted, we also need the declared return type available at this position. A similar threading through of information will be necessary for parameter types, but is out of scope for this diff.

Reviewed By: Matt-Schellhas

Differential Revision: D15365111

fbshipit-source-id: ca8f2d8b03b4d176b551334cf37c285aa9349127