Force the symbol index test to sqlite
Summary:
As reported by Oguz Ulgen this particular test was failing sporadically:

`buck test mode/dbgo hphp/hack/test/integration:run_symbol_index_test`

Root cause of the failure is that this test works correctly for `SqliteIndex` mode, but does not work correctly in `TrieIndex` mode.  Apparently, the TrieIndex thinks that "UsesA" is a function rather than a class.

The fix is to force the test into Sqlite mode, which is what we want to test anyway.

Reviewed By: oulgen

Differential Revision: D15504370

fbshipit-source-id: da1974b47c83a25712b1dc8552273cb48ec8bddc