Re-Use Zstd Contexts
Summary:
Approximately 70% of Zstd CPU in HHVM is spent allocating and freeing
`ZSTD_CCtx`s. This diff adds a simple pool of contexts that can be reused.
This implementation, although it's a simple lock-guarded vector, has worked
quite well for Managed Compression and has never been observed to be a
point of significant contention.

Reviewed By: terrelln

Differential Revision: D15398650

fbshipit-source-id: 317b82bf55f40d555f43fd2081c948cf79861ed1