Remove dead part of prepareArrayArgs()'s API
Summary:
This function always returns true, so change it to return void and kill all the
dead branches at its both call sites. Simplify doInitArgs() further as it always
returns false.

It also takes a `retval` argument, which if non-nullptr, it writes KindOfNull to
in certain exceptional situations. One caller always passes nullptr, so it does
nothing there. The other caller passes a local variable that is never read, so
it's not much useful either. So, kill the arg.

Reviewed By: ricklavoie

Differential Revision: D15527352

fbshipit-source-id: 11c536bee4207867cc9b79007f32a6c564fad3db