Parse raw address string in VariableUnserializer::matchString
Summary: For APC serialization, if the string is a static string, instead of recording the contents of the string, we just record the raw memory address of this string. This diff makes it so that `matchString` function can read this format, hence preventing it from punting to the slower path.

Reviewed By: markw65

Differential Revision: D15529273

fbshipit-source-id: 21d50d4e91e514ab24ae9849fd58e3ce9c04a6b2