Split FPush/FCall opcode checks to legacy, new and both
Summary:
Split isFPush(), isFCallStar() helpers into isLegacyFPush(), isLegacyFCall(),
hasFPushEffects() and hasFCallEffects(), plus add a new helper isNewFCall()
that will be used for the new merged FPush/FCall opcodes.

Audit all places these helpers are used and all places directly checking for
OpFCall and Op::FCall and determine which helper is appropriate for which
place:

- FPI region tracking uses isLegacy*() helpers
- control flow effects use hasFCallEffects()
- stack bound checking uses hasFCallEffects()
- inlining logic uses has*Effects()

This diff is a noop, as isLegacyFoo() and hasFooEffects() are equivalent as of
now, but prepares ground for FCallCtor.

Reviewed By: alexeyt

Differential Revision: D15459372

fbshipit-source-id: 1a0243101e4d1e5e01a51e2ccfe8a53a65785111