Back out "[hhvm] Make Shapes::idx an optimized builtin"
Summary:
This diff causes an significant regression when jumpstart is disabled.

We're not sure of the cause of the regression. This change doesn't touch any
TargetProfile logic directly, but it results in many new ArrayOffsetProfiles
being created in all the places that Shapes::idx gets inlined.

In theory, new profiling code shouldn't affect performance post retranslateAll,
but in practice, evidently there is some effect.

Original commit changeset: 6995df09a01e

Reviewed By: mofarrell

Differential Revision: D15546271

fbshipit-source-id: 2f37046a986578c0d41717281ecd3b7eefed8a66