Move the FPI stack region formation fix to emitFCall
Summary:
Add a bit to IRGS indicating whether we are forming a region. Use it from
ldPreLiveFunc() used by emitFCall() to skip loading Func* from FPI stack if we
are forming a region. Also drop a pre-optimization for LdARFuncPtr, since it
was doing the same thing (and also existed long before, so not sure how exactly
the new code triggered the issue).

Replaces D15493575's logic in an equivalent way, but adds infra for the new
FCall* opcodes and makes it more obvious what needs to be done there.

Reviewed By: paulbiss

Differential Revision: D15549918

fbshipit-source-id: 795a8c18424cd21d27d7558b2635fe7647d88023