Add NewObjRD as an optimizations for ReifiedName + NewObj
Summary:
This diff adds a new bytecode to optimize the combination of `ReifiedName` + `NewObj` since for reified generics we do know the name of the class statically but using `NewObj` was preventing us from using this static knowledge.
According to strobelight data, `ReifiedName` is a very expensive instruction. The goal is to remove this instruction completely.
Prior to this diff, the only remaining use sites of `ReifiedName` were `new C<int>()` and `f<int>()`. This diff removes the class instantiation. On a follow up diff, I'll remove function call. Ultimately, this will lead to deleting `ReifiedName` and all the accompanying complex mechanism.

Reviewed By: jano

Differential Revision: D15540960

fbshipit-source-id: e5c55f91a3f526113375502b9605d017e64a737e