Simplify coercion and make it match the typing rules
Summary: This bans coercion to generic types that are bounded below by dynamic, requiring an explicit `as dynamic` assertion. It also removes the explicit ban on coercion to union types, which used to be there for unresolved types in the legacy type system. I updated the docs to remove the transitivity rule which was enabled by the generic rule.

Reviewed By: dlreeves

Differential Revision: D15542522

fbshipit-source-id: 2a542b07fff0778e0cf161c0d91117a883132d39