Remove Awaitable coercion rule, unwrap Awaitable coerced type
Summary:
This diff removes the coercion rule `T1 ~> T2 |- Awaitable<T1> ~> Awaitable<T2>`. Instead, when considering a `Return (Some e)` node in the Aast, we unwrap the Awaitable before passing it to coercion.

This also fixes a corner case in the typechecker that allowed coercion from Awaitable<T> to Awaitable<dynamic> in a non-async function, where there's no enforcement of the inner type of the Awaitable (even though in this case the inner type of the target is dynamic).

Reviewed By: dlreeves

Differential Revision: D15549422

fbshipit-source-id: a8a0635b5a20980606d9e50268e07f97951648d9