Remove the Awaitable wrapping of empty return
Summary: D15549422 covered the `return expr;` case. This covers the `return;` case. It unwraps the declared return type for async, generators, and async generators. For generators, the only way to get a returned value is by the `getReturn()` method which Hack doesn't expose in the Generator API. Therefore, it doesn't make sense to call `Typing_return.implicit_return` (which does coercion) in this case.

Reviewed By: dlreeves

Differential Revision: D15551647

fbshipit-source-id: e3397a89c6110d486ded930f732599d4dcff3042