local allocator for thread-local data
Summary:
The local allocator is designed to hold data that is local to the worker thread, such as request heap, data allocated there are generally expected to be hot and short lived.  If the allocated data are also accessed from other threads, they should not be allocated in the local allocator (and that's why we need a separate arena rather than the per-thread default arena).
It is also useful in some extensions where using request allocator gets awkward.

Reviewed By: markw65

Differential Revision: D15545888

fbshipit-source-id: dae6ff699fa4cb15626e113016033b2659c6846e