Sync FB / OSS build on build_id
Summary:
- FB side: build_id is regenerated everytime a dependency of hh_single_compile
  is modified
- dune/OSS side: build_id is regenerated once per 'clean' build

This patch makes that with dune/OSS, the build_id is always regenerated (cheap
operation, since if the build-id.c file is the same, it's compilation is not
necessary and dune doesn't rebuild it) so it is in sync with FB when it matters.
(Also unify the hh vs hackc prefix as OSS doesn't care about it being exactly
hh).

Reviewed By: fredemmott

Differential Revision: D15536225

fbshipit-source-id: d27d85072aff455563acd32f41c88d442aaabb8d