Switch mixed-process test to full ocaml unit test
Summary:
As reported by Tatiana Racheva, the test I wrote for validating the symbol index service relies on launching HH.  Because it's an integration test, it depends on consistent behavior of watchman + HH Server + HH Client.

In experience it was flaky.

So, let's rewrite it to simply execute the same statements we want to test locally.

Reviewed By: 2BitSalute

Differential Revision: D15566608

fbshipit-source-id: 2e83486f8f5b2077e47d559887086f8f48022240