Fix unsoundness in is/instanceof on variable of generic parameter type
Summary:
When checking refinements, we add bounds on type parameters that necessarily hold as a result of witnessing a type at runtime. For example:
```
class C<T> { }
class D extends C<B> { }
function foo<T>(C<T> $c, T $t): B {
  if ($c is D) {
  // Here, we know that T=B
    return $t;
  }
  ...
}
```
We have two types in our hand: the static type of the scrutinee (here, `C<T>`), and the witnessed type (here, `D`). The addition of bounds is only sound when the type of the scrutinee is a concrete classish type, or if it has a concrete supertype, as in this variation:
```
class C<T> { }
class D extends C<B> { }
function foo<T, Tc as C<T>>(Tc $c, T $t): B {
  if ($c is D) {
    return $t;
  }
  ...
}
```
The existing code is unsound because it will attempt to add bounds to the environment if the type of the scrutinee is a generic type parameter. Instead, we should chase all its concrete supertypes.

Reviewed By: manzyuk

Differential Revision: D15536997

fbshipit-source-id: 83a5c7d3bec23d27ba025b12fcff058df9067c79