Factor out common profiling code
Summary:
This diff pulls out the basic "first n distinct elements" profiling logic into FrequencyProfile and uses it in ArrayOffsetProfile and CallTargetProfile.

It also simplifies this structure a little bit. In the old implementation, we need to check whether the profile is initialized (i.e. has processed its first input) because the struct starts out 0-initialized in RDS. Rather than doing that, I've opted to make FrequencyProfile use 0s as its "unoccupied entry" value and to add 1 to the offsets being profiled in ArrayOffsetProfile and the funcIDs in CallTargetProfile.

This diff fixes a bug in ArrayOffsetProfile that fails to correctly merge two profiles. This bug is asymptomatic, because when the profile hits, it tends to be *very* (>90%) concentrated on a single offset.

Reviewed By: mofarrell

Differential Revision: D15502373

fbshipit-source-id: d6aac1548f9ff1180031ad863626476f4778cfe0