Take account of address changes in store-elim
Summary:
Alias analysis assigns a unique aloc to each (suitable) PureStore's
address. store-elim then assumes that each aloc represents a single
memory location.

This is always safe for stack locations, or locals, but for locations
on heap allocated things, its only safe as long as the address doesn't
change.

Since we use ssa form, the def of the address dominates the store, so
its generally ok even for heap allocated memory; but in a loop, its
possible that the store is partially available at the def of its
address; if its also partially anticipated, we could end up assuming
that stores from one iteration are partially redundant wrt stores from
the next, even though the next iteration is actually storing to a
different location in memory.

We can solve this by giving stores a "dependent" address, and adding
fake loads at the point where the dependent address is defined.

Reviewed By: ricklavoie

Differential Revision: D15593915

fbshipit-source-id: ee53fc3dc19f12928f551ca3f4b1a16de4f8aef5