Add ::getUncompressedByteCount()
Summary:
gzip encoding can consist of multiple consecutive streams. When we use gunzip or gzopen()/gzread() these streams are automatically concatenated on the output:
https://github.com/madler/zlib/blob/master/gzread.c#L215
(when a stream ends, said functions will automatically look for a new stream)

These files don't function very well with the current inflateChunk API since trailing bytes are not exposed. The right behavior in these cases is that, when we close a stream (due to Z_STREAM_END), the caller should check if there's any undecompressed bytes from the last iteration, and reinitialize the decompressor if necessary.

There's a new test that shows how to decompress such files.

Reviewed By: jano

Differential Revision: D15539582

fbshipit-source-id: 004131e446158014190678fef0b4f9a20ae0a80a