Process environment should be controllable when working directory is specified
Summary:
Simplify and make more regular the Process module's exec methods' handling of the environment by introducing the environment type which can be:

* Default: the current process's environment
* Empty: this was and remains the behavior if a working directory is specified
* Augment: use the current environment and add to it
* Replace: not currently used, adding for completeness - works as if the empty environment was augmented instead of the current one

My goal was to refactor and add the ability to specify the environment when setting the working directory, which was not previously possible. Otherwise, I intended to keep the original behavior the same.

Differential Revision: D15553222

fbshipit-source-id: 4d787f8928745676aee9ec7e7c7fcb276e383ca0