The Process's exec-with-cwd behavior needs explanation
Summary:
What the comment says I learned the hard way. The API suggests that the working directory is changed before the target process is executed, but that's not actually what happens. Instead, we respawn our own process as a daemon with the target working directory, then have it execute the target process, redirect the output, and stop the daemon.

If you're creating a new executable and you end up calling this API without having called Daemon.check_entry_point () first,  the results will be surprising!

Differential Revision: D15550611

fbshipit-source-id: 52d220c2ef2572ed8ee3f1ca21d1b6d402816faa