Allow DCE for array lookup ops
Summary:
AKExistsArr, ArrayIdx, and ArrayIsset are currently marked as opcodeMayRaise, which means DCE can't remove them. These ops used to raise when certain array compatibility flags were set. However, we now check these flags in irgen, and most of the time we can guarantee that we won't raise based on static type knowledge. When we do need to raise a notice, we emit separate IR ops like RaiseArrayIndexNotice and its friends.

In this diff, we we remove the taken edge from these ops and allow them to DCE.

Reviewed By: markw65

Differential Revision: D15596391

fbshipit-source-id: 77a91899076b118d4696ad7d1b4614fea50e5d8e