Split LogArrayProvenance
Summary:
Two flags:

ArrayProvenance controls instrumentation of array provenance, period.

LogArrayProvenance controls logging.

There's some funny business where I turn off the instrumentation if was enabled
in the repo but we start HHVM without the logging enabled; or, if it wasn't
enabled when we built the repo.

Either way, if we're not loading from a repo, the options in the hdf are
respected without question :)

Reviewed By: mxw

Differential Revision: D15545217

fbshipit-source-id: 139a00ef48acdedb93c763d1e5ffebf9820a1620