Call NvGetStr() from ExistsStr()
Summary:
HPHP::MixedArray::ExistsStr() called frequently and has measurable amount of cycle/instruction samples. Currently the function calls to C++ search function.
But in this case the NvGetStr() assembly function can be used to improve the performance of the calls.

SL shows slight improvement in the interesting counters. The expectation to get up to 0.2% in production.
https://our.intern.facebook.com/intern/servicelab/33689710/
https://our.intern.facebook.com/intern/servicelab/33689709/
https://our.intern.facebook.com/intern/servicelab/33689708/
https://our.intern.facebook.com/intern/servicelab/33689707/
https://our.intern.facebook.com/intern/servicelab/33689706/
https://our.intern.facebook.com/intern/servicelab/33689705/
https://our.intern.facebook.com/intern/servicelab/33689704/
https://our.intern.facebook.com/intern/servicelab/33689703/

Reviewed By: binliu19, alexeyt, markw65

Differential Revision: D15556118

fbshipit-source-id: 9565b6abd76677da04affd09d02b81b518cf31b4