Unconditionally raise notice on inc/dec on non-numeric strings
Summary:
EnableHipHopSyntax option is being defaulted to true and removed.

One of the behaviors it controls is whether to to issue a notice when trying
to increment or decrement non-numeric strings. Do it unconditionally and update
some and remove all other affected unit tests, as ability to increment or
decrement non-nums will be eventually removed.

Reviewed By: kmeht

Differential Revision: D15601880

fbshipit-source-id: 69a30d2212a2a56c6ea517057becdf17afa67db9