Fix a bug with clsrefgetts
Summary: If the input type structure to `ClsRefGetTS` is not static and it does not exist on the named entity table, we'll turn it into static to use it as a key but that results in the pointer being changed and the old pointer being invalid. Let's use the new pointer.

Reviewed By: ricklavoie

Differential Revision: D15610982

fbshipit-source-id: 3d0b32a5cef58195c04a82e3f256c0213fca26e6