Add symbol capture for constants and typedefs
Summary:
After review of the rollout of sqlite index, we are comparing logs from users who have searched for things not found.

I discovered two elements that were not being captured that are searched for: typedefs and constants.

We'll also have to figure out a solution for XHP classes, but that won't be in this small diff.

Additionally, Microsoft seems to have added a few new kinds to its LSP definition, so we'll update them.  Thankfully, there is now a defined kind for "constant!"  Which will make autocomplete slightly less imprecise.

https://microsoft.github.io/language-server-protocol/specification
{F160121206}

I also took this opportunity to remove a few unnecessary calls to core_kernel :)

Reviewed By: ljw1004

Differential Revision: D15620226

fbshipit-source-id: f310f440c3a99968a7393923068d6d8083de9296