Remove the special case of typing null checks
Summary: Previously, in strict mode, we would attach to `null` in the expression `$x === null` the same type in the TAST as to `$x` -- even if the latter is non-nullable!  I don't think we need this special case anymore.  It was introduced so that we could enforce that non-nullable values are not compared to `null`, but that check has been moved out into a TAST check (`Tautology_check`) as it is not really a type error.

Reviewed By: kmeht

Differential Revision: D15625032

fbshipit-source-id: cadb60db1fc6b58b1a381cc7ab99c13c845ca6a9