Limit array_multisort() to 9 arguments
Summary:
array_multisort() is the last builtin accepting variadic arguments by
reference. Let's restrict it to a reasonable maximum number of arguments that
allow removal of support for variadic by ref.

There's a minor semantic difference, where passing a null would be previously
interpreted as SORT_REGULAR. This won't cause any issues, because SORT_REGULAR
is the default, so it would change outcome only if passed together with other
previously specified SORT_ option, which almost certainly indicates a bug. The
most likely source of such a bug would be something like, where it is clearly
undesirable to interpret it as SORT_REGULAR:

  $arr1 = array(...);
  $sort_string = SORT_STRING;
  $arr2 = get_null_instead_of_array();
  array_multisort(&$arr1, &$sort_string, &$arr2);

Reviewed By: paulbiss

Differential Revision: D15671931

fbshipit-source-id: 3406e579fd5353628f5d865fed108cb5f3aa7121