Allow nonexistent includes within the watchman root
Summary:
Let's assume the watch root is `/path/to/repo` and the path to the `.flowconfig` is `/path/to/repo/myproject/src`.

## Unchanged behaviors

Here's what happens when Flow sees these in the `[include]` section while using Watchman as a file watcher. These are not changing

* `path/to/existent` - We would watch `/path/to/repo/myproject/src/path/to/existent`
* `../../../` - Watchman would error `/path/to` is outside of the watch root and Flow would exit
* `../../../other_repo` - Flow would exit, since it only supports 1 watch root at a time.

## Changing behavior

```
[include]
path/to/nonexistant
```

Previously, Watchman would error and Flow would crash. Now we correctly watch `/path/to/repo/myproject/src/path/to/existent` in case it shows up later.

## More details

Flow uses `watchman watch-project <PATH>` to figure out which watchman project `<PATH>` is a part of. `watchman watch-project` returns the *watch root*. If `<PATH>` doesn't exist on disk, then `watchman watch-project` will error. This can be triggered by adding a non-existent path to the `[include]` section of the `.flowconfig`.

This diff makes a slight change to the error behavior. It allows non-existent paths in the `[include]` section if and only if they start with the watch root (e.g if the watch root is `/path/to/repo` then `/path/to/repo/foo` is allowed but `/tmp/bar` still errors).

Reviewed By: mroch

Differential Revision: D15665329

fbshipit-source-id: 2610ed1c74a08398190c3de0b5f566d0b18b9eda