Keep real recheck stats around
Summary:
The current recheck stats record is overwritten every time we, for example, run hh. This makes it difficult to use for saved state heuristics because it could be inaccurate.

In this change, I'm adding another stats record that would only be set in the same path where we log the recheck ID/recheck stats.

Our existing saved state and IDE tests rely on the existing recheck stats behavior, since it's convenient for them to check that the rechecked count is 0 if it is, indeed, a no-op. Therefore, I'm not touching anything about how the existing field is handled and simply adding a new one.

Differential Revision: D15599010

fbshipit-source-id: 63f03a854c906ff893aa9edc55be8d402d74af93