Fix issue with PopU2 in dce
Summary:
In some cases, dce is all or nothing. Eg an AddElemC pops 3 things
from the stack, so we mark them as "linked". If we decide not to kill
any of the producers, we won't kill the AddElemC.

Unfortunately, PopU2 can put things in between those elements (since
dce runs backwards, a "pop" actually inserts something onto the
stack).

The result is that if, eg one of the 3 cells was generated by an
FCallBuiltin, the PopU2 inserts an element, the FCallBuiltin says its
not dead, which then marks the AddElemC and the PopU2 as live, rather
than the AddElemC and the thing under the PopU2.

Fix it by having the PopU2 mark a chain of elements as live if it
inserts itself in the middle. This generally wont hurt, because the
PopU2 will be killed, we'll reprocess the block, and now the links
will work as intended.

Reviewed By: ricklavoie

Differential Revision: D15683441

fbshipit-source-id: 21e7dae7359279634dcdf4a4b169281f201f1157