Add per-SrcKey counters and only emit live/profile translations beyond a threshold
Summary:
Some functions, despite being hot enough to be JITed, may still contain cold
portions that are not hot enought to justify JITing them.  This diff adds
separate counters for each SrcKey and only triggers translation of SrcKeys after
a threshold is hit.  The per-function counters are still used, so that any
SrcKey for a function is only JITed after the function threshold is hit.  This
aims at preserving the locality of the hot translations for each function.
This diff leaves the new mechanism disabled by default, by having the new
threshold set to 0.

Reviewed By: binliu19

Differential Revision: D15592828

fbshipit-source-id: ec820a054907ba7f46a85041c840d4309b86e52d